requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.0.74
rasterio==1.3.8
geopandas==0.13.2
shapely==2.0.1
folium==0.14.0
matplotlib==3.7.1
plotly==5.15.0
scikit-learn==1.3.0
scipy==1.10.1
pandas==2.0.2
fastapi==0.100.0
uvicorn==0.23.2
pydantic==2.0.3
pyyaml==6.0
pillow==10.0.0
seaborn==0.12.2
gdal==3.7.2